begins commenting, cleans up more things
