+++ crates/instantiator/README.md
@@ -1,3 +1,8 @@
+# Metaverse Instantiator 
+
+this project allows for spawning async metaverse servers, for use in testing other components of the project. 
+see the test dir for use 
+
 # getting started 
 ## dependencies 
 - mono-core
